event-driven/integration-service/requirements.txt
Flask==3.0.0
Flask-RESTx==1.3.0
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
structlog==23.2.0